hotfix(Node): correcao

+++ Includes/Classes/Node/Node.hpp
@@ -5,7 +5,7 @@ class Node {
 private:
   int row{0};
   int col{0};
-  int value{0};
+  double value{0};
   Node *down{nullptr};
   Node *next{nullptr};
